+++ packages/wonder-blocks-clickable/src/components/clickable.tsx
@@ -113,37 +113,6 @@ type CommonProps =
 
 type Props =
     | (CommonProps & {
-          target?: never;
-          beforeNav?: never;
-          safeWithNav?: never;
-      })
-    | (CommonProps & {
-          href: string;
-
-          /**
-           * A target destination window for a link to open in.
-           */
-          target?: "_blank";
-
-          beforeNav?: never;
-          safeWithNav?: never;
-      })
-    | (CommonProps & {
-          href: string;
-
-          /**
-           * Run async code before navigating. If the promise returned rejects then
-           * navigation will not occur.
-           *
-           * If both safeWithNav and beforeNav are provided, beforeNav will be run
-           * first and safeWithNav will only be run if beforeNav does not reject.
-           */
-          beforeNav: () => Promise<unknown>;
-
-          safeWithNav?: never;
-          target?: never;
-      })
-    | (CommonProps & {
           href: string;
 
           /**
@@ -152,7 +121,7 @@ type Props =
            * settled before the navigation will occur. Errors are ignored so that
            * navigation is guaranteed to succeed.
            */
-          safeWithNav: () => Promise<unknown>;
+          safeWithNav?: () => Promise<unknown>;
 
           /**
            * A target destination window for a link to open in.
@@ -162,7 +131,7 @@ type Props =
           beforeNav?: never;
       })
     | (CommonProps & {
-          href: string;
+          href?: string;
 
           /**
            * Run async code before navigating. If the promise returned rejects then
@@ -171,7 +140,7 @@ type Props =
            * If both safeWithNav and beforeNav are provided, beforeNav will be run
            * first and safeWithNav will only be run if beforeNav does not reject.
            */
-          beforeNav: () => Promise<unknown>;
+          beforeNav?: () => Promise<unknown>;
 
           /**
            * Run async code in the background while client-side navigating. If the
@@ -179,7 +148,7 @@ type Props =
            * settled before the navigation will occur. Errors are ignored so that
            * navigation is guaranteed to succeed.
            */
-          safeWithNav: () => Promise<unknown>;
+          safeWithNav?: () => Promise<unknown>;
 
           target?: never;
       });