+++ packages/wonder-blocks-data/src/util/status.ts
@@ -4,6 +4,10 @@ const loadingStatus = Object.freeze({
     status: "loading",
 });
 
+const noDataStatus = Object.freeze({
+    status: "no-data",
+});
+
 const abortedStatus = Object.freeze({
     status: "aborted",
 });
@@ -14,6 +18,8 @@ const abortedStatus = Object.freeze({
 export const Status = Object.freeze({
     loading: <TData extends ValidCacheData = ValidCacheData>(): Result<TData> =>
         loadingStatus,
+    noData: <TData extends ValidCacheData = ValidCacheData>(): Result<TData> =>
+        noDataStatus,
     aborted: <TData extends ValidCacheData = ValidCacheData>(): Result<TData> =>
         abortedStatus,
     success: <TData extends ValidCacheData>(data: TData): Result<TData> => ({