+++ packages/wonder-blocks-clickable/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @khanacademy/wonder-blocks-clickable
 
+## 4.0.2
+
+### Patch Changes
+
+-   8dc4a5a3: Allow 'beforeNav' and 'safeWithNav' to be used without an 'href'
+-   Updated dependencies [1920feb8]
+    -   @khanacademy/wonder-blocks-core@6.0.1
+
+## 4.0.1
+
+### Patch Changes
+
+-   f230b267: Allow 'safeWithNav()' to be used on its own without 'beforeNav()' in prop types
+
 ## 4.0.0
 
 ### Major Changes