+++ packages/wonder-blocks-link/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/wonder-blocks-link
 
+## 5.0.2
+
+### Patch Changes
+
+-   Updated dependencies [8dc4a5a3]
+-   Updated dependencies [1920feb8]
+    -   @khanacademy/wonder-blocks-clickable@4.0.2
+    -   @khanacademy/wonder-blocks-core@6.0.1
+    -   @khanacademy/wonder-blocks-icon@2.1.2
+
+## 5.0.1
+
+### Patch Changes
+
+-   Updated dependencies [f230b267]
+    -   @khanacademy/wonder-blocks-clickable@4.0.1
+
 ## 5.0.0
 
 ### Major Changes